+++ cache/realm/src/main/java/io/github/zmunm/insight/cache/table/TableGame.kt
@@ -5,7 +5,7 @@ import io.realm.annotations.PrimaryKey
 import java.util.Date
 
 internal open class TableGame(
-    @PrimaryKey var id: Int = 0,
+    @PrimaryKey var id: Long = 0,
     var name: String = "",
     var backgroundImage: String = "",
     var addedAt: Date = Date(),
